# replace string slices with startsWith comparison