@guybedford 2020-10-03
I must admit I am somewhat shocked V8 doesn't just do this optimization internally, but the PR looks great thank you. I will update the benchmarks now.