@guybedford 2020-10-03
Yeah I'm also seeing a ~20% improvement here, benchmarks updated in https://github.com/guybedford/cjs-module-lexer/commit/9a551db2f3de853e005d1d672a0bcb575d6b6194.

Note that the huge benefit of Wasm is not the warm speed but the cold speed because the compilation doesn't need to be done JIT which saves ~ 250ms of compile time V8 otherwise has to do.